+++ packages/yourloops/components/patient-data/patient-data.tsx
@@ -158,7 +158,6 @@ export const PatientData: FunctionComponent = () => {
                         bgPrefs={bgPrefs}
                         medicalDataService={medicalData}
                         patient={patient}
-                        timePrefs={timePrefs}
                         loading={refreshingData}
                       />
                     }