+++ CHANGELOG.md
@@ -3,11 +3,16 @@ Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
 ## 3.0.6-beta - 2023-11-10
+### Changed
+- YLP-2574 Create tool tips independently of the name of the columns
+
 ### Engineering use
 - YLP-2476 Use tide-whisperer v2: route /all.
 - YLP-2696 Fix some Sonarcloud bugs and code smells
+- YLP-2704 Remove settings logic from viz
+- YLP-2726 Cleanup tideline/viz packages
 
-## 3.0.5-beta - 2023-07-12
+## 3.0.5 - 2023-11-02
 ### Added
 - YLP-2292 Display rescue carbs and manual bolus in the trends view for DBLG1 patient
 - YLP-2299 Display DBLG1 events in Daily view